commits: 1. add async fusion algo 2. fix motion manager service 3. add new api for dag streaming
