+++ cmake-build-debug/CMakeFiles/clion-log.txt
@@ -1,0 +1,1 @@
+CMakeLists.txt not found in /home/w/dev/code/apollo
