+++ modules/perception/common/perception_gflags.cc
@@ -120,3 +120,8 @@ DEFINE_bool(capture_screen, false, "");
 DEFINE_string(screen_output_dir, "./", "");
 
 DEFINE_string(frame_visualizer, "GLFusionVisualizer", "");
+
+DEFINE_double(car_length, 3.564, "car_length");
+DEFINE_double(car_width, 1.620, "car_width");
+
+DEFINE_bool(has_lidar, true, "if perception has lidar sensor");
