+++ modules/perception/conf/perception_lowcost.conf
@@ -199,3 +199,5 @@
 # type: string
 # default: CCLanePostProcesser CCLineProcesser
 --onboard_lane_post_processer=CCLanePostProcesser
+
+--has_lidar=false
