+++ modules/perception/obstacle/fusion/probabilistic_fusion/BUILD
@@ -17,6 +17,7 @@ cc_library(
         "pbf_track_manager.cc",
         "pbf_track_object_distance.cc",
         "probabilistic_fusion.cc",
+        "pbf_imf_fusion.cc",
     ],
     hdrs = [
         "pbf_base_motion_fusion.h",
@@ -31,6 +32,7 @@ cc_library(
         "pbf_track_manager.h",
         "pbf_track_object_distance.h",
         "probabilistic_fusion.h",
+        "pbf_imf_fusion.h",
     ],
     data = [
         "//modules/perception/conf:perception_config",
@@ -38,6 +40,7 @@ cc_library(
     deps = [
         "//modules/common",
         "//modules/common:log",
+        "//modules/perception/common",
         "//modules/perception/lib/config_manager",
         "//modules/perception/obstacle/base",
         "//modules/perception/obstacle/common",
