+++ modules/perception/obstacle/fusion/probabilistic_fusion/pbf_imf_fusion.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  *****************************************************************************/
 
-#include "modules/perception/obstacle/fusion/imf_fusion/imf_fusion.h"
+#include "modules/perception/obstacle/fusion/probabilistic_fusion/pbf_imf_fusion.h" // NOLINT
 #include "modules/common/log.h"
 #include "modules/perception/obstacle/base/types.h"
