+++ modules/perception/onboard/dag_streaming.cc
@@ -40,6 +40,9 @@ DEFINE_int32(max_allowed_congestion_value, 0,
 DEFINE_bool(enable_timing_remove_stale_data, true,
             "whether timing clean shared data");
 
+SubnodeMap DAGStreaming::subnode_map_;
+std::unordered_map<std::string, SubnodeID> DAGStreaming::subnode_name_map_;
+
 DAGStreaming::DAGStreaming()
     : Thread(true, "DAGStreamingThread"),
       inited_(false),
@@ -171,7 +174,7 @@ bool DAGStreaming::InitSubnodes(const DAGConfig& dag_config) {
       return false;
     }
     subnode_map_.emplace(subnode_id, std::unique_ptr<Subnode>(inst));
-
+    subnode_name_map_[subnode_config.name()] = subnode_id;
     AINFO << "Init subnode succ. " << inst->DebugString();
   }
 
@@ -223,6 +226,15 @@ void DAGStreamingMonitor::Run() {
     }
     sleep(1);
   }
+}
+
+Subnode* DAGStreaming::GetSubnodeByName(std::string name) {
+  std::unordered_map<std::string, SubnodeID>::iterator iter
+          = subnode_name_map_.find(name);
+  if (iter!= subnode_name_map_.end()) {
+    return subnode_map_[iter->second].get();
+  }
+  return nullptr;
 }
 
 }  // namespace perception
